pyproject.toml
[tool.poetry]
name = "groceries-etl"
version = "1.0.0"
description = "Weekly grocery deals ETL pipeline"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "groceries", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.29.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
click = "^8.1.0"
aiofiles = "^23.2.0"
httpx = "^0.25.0"
beautifulsoup4 = "^4.12.0"
lxml = "^5.0.0"
playwright = "^1.40.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
mypy = "^1.7.0"
ruff = "^0.1.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008", "B006"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

